split main itest into multiple files
