+++ tests/integration/test_charm.py
@@ -25,44 +25,6 @@ prometheus_resources = {"prometheus-image": oci_image("./metadata.yaml", "promet
 
 
 @pytest.mark.abort_on_fail
-async def test_build_and_deploy_with_alternative_images(ops_test, prometheus_charm):
-    """Test that the Prometheus charm can be deployed successfully."""
-    app_name = "prometheus-ubuntu"
-
-    await ops_test.model.deploy(
-        prometheus_charm, resources=prometheus_resources, application_name=app_name
-    )
-    await ops_test.model.wait_for_idle(apps=[app_name], status="active")
-    await ops_test.model.block_until(lambda: len(ops_test.model.applications[app_name].units) > 0)
-
-    assert ops_test.model.applications[app_name].units[0].workload_status == "active"
-
-    await check_prometheus_is_ready(ops_test, app_name, 0)
-
-    await ops_test.model.applications[app_name].remove()
-    await ops_test.model.block_until(lambda: app_name not in ops_test.model.applications)
-    await ops_test.model.reset()
-
-
-@pytest.mark.abort_on_fail
-async def test_build_and_deploy_prometheus_tester(ops_test, prometheus_tester_charm):
-    """Test that Prometheus tester charm can be deployed successfully."""
-    app_name = "prometheus-tester"
-
-    await ops_test.model.deploy(
-        prometheus_tester_charm, resources=tester_resources, application_name=app_name
-    )
-    await ops_test.model.wait_for_idle(apps=[app_name], status="active")
-    await ops_test.model.block_until(lambda: len(ops_test.model.applications[app_name].units) > 0)
-
-    assert ops_test.model.applications[app_name].units[0].workload_status == "active"
-
-    await ops_test.model.applications[app_name].remove()
-    await ops_test.model.block_until(lambda: app_name not in ops_test.model.applications)
-    await ops_test.model.reset()
-
-
-@pytest.mark.abort_on_fail
 async def test_prometheus_scrape_relation_with_prometheus_tester(
     ops_test, prometheus_charm, prometheus_tester_charm
 ):
